requirements.txt
flask==1.1.2
pillow==8.3.1
tensorflow==2.5.0
requests==2.25.1 # Used in testing.py